refactor record trace mapping helpers
